move trendingworks to home folder
