+++ src/components/Home/TrendingWorks/TrendingWorks.styles.ts
