+++ src/components/Home/TrendingWorks/TrendingWorks.types.ts
